requirements.txt
affine==2.4.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.7.4
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
exceptiongroup==1.2.2
executing==2.0.1
fiona==1.9.6
fonttools==4.53.1
geopandas==0.14.4
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyogrio==0.9.0
pyparsing==3.1.4
pyproj==3.6.1
pyspatialml==0.22.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.2.0
rasterio==1.3.10
rioxarray==0.17.0
scikit-learn==1.5.1
scipy==1.14.1
shapely==2.0.6
six==1.16.0
snuggs==1.4.7
sqlparse==0.5.1
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13
xarray==2024.7.0
